feat(sentry): move sentry to util
